Feat: 수정페이지 유효성 검증 추가, 프로필사진 변경하지 않았는데 기존 사진파일 삭제되는 오류 수정

+++ app-backend/src/main/java/com/mmk/controller/UserController.java
@@ -160,13 +160,15 @@ public class UserController {
         try {
             UserDTO uv = new ObjectMapper().readValue(userInfoJson, UserDTO.class);
 
-            if (uv.getProfile() != null && !uv.getProfile().isEmpty()) {
-                String existingProfilePath = System.getProperty("user.dir") + "/app-backend/src/main/resources/static"
-                        + uv.getProfile();
-                File existingFile = new File(existingProfilePath);
-
-                if (existingFile.exists()) {
-                    existingFile.delete();
+            if (profileImage != null && !profileImage.isEmpty()) {
+                if (uv.getProfile() != null && !uv.getProfile().isEmpty()) {
+                    String existingProfilePath = System.getProperty("user.dir") + "/app-backend/src/main/resources/static"
+                            + uv.getProfile();
+                    File existingFile = new File(existingProfilePath);
+    
+                    if (existingFile.exists()) {
+                        existingFile.delete();
+                    }
                 }
             }
 
@@ -209,7 +211,6 @@ public class UserController {
             }
 
             userService.updateUser(uv);
-            // userRepository.save(uv);
 
             ApiResponse<UserDTO> response = new ApiResponse<>(201, "회원 정보 수정 성공", uv);
             return ResponseEntity.ok(response);
